- Added sourcetrail configuration
- Updated comments
- Added doxygen
