+++ .gitignore
@@ -2,6 +2,8 @@
 
 /CMakeFiles
 
+doc/doxygen
+
 CMakeCache.txt
 
 /.vscode/*
