+++ libraries/groups.h
@@ -1,0 +1,18 @@
+
+/**
+ * @defgroup device_drivers Device Drivers
+ * 
+ * This group contains device drivers of used mcu
+ * 
+ * @{
+ * @}
+ */
+
+/**
+ * @defgroup peripheral_drivers Peripheral drivers
+ * 
+ * This group contains drivers for devices connected to board
+ * 
+ * @{
+ * @}
+ */
