+++ platforms/mcu/EngModel/inc/mcu/config.hpp
@@ -20,7 +20,7 @@
 /******************************************************************************\
  *                             Includes
 \******************************************************************************/
-#include "stm32f1xx_hal.h"
+#include "TargetsDefines.hpp"
 
 /******************************************************************************\
  *                              Code
