+++ sourcetrail/Adapter_Board.srctrlprj
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<config>
+    <source_groups>
+        <source_group_2fd9e6cd-1a40-4096-a649-56ef9c17a8c6>
+            <build_file_path>
+                <compilation_db_path>../build/compile_commands.json</compilation_db_path>
+            </build_file_path>
+            <compiler_flags>
+                <compiler_flag>--target=arm-unknown-gnueabi</compiler_flag>
+            </compiler_flags>
+            <header_search_paths>
+                <header_search_path>../libraries/external/STM32CubeF1/Drivers/CMSIS/Core/Include</header_search_path>
+                <header_search_path>C:/Program Files (x86)/GNU Tools Arm Embedded/9 2019-q4-major/arm-none-eabi/include</header_search_path>
+                <header_search_path>C:/Program Files (x86)/GNU Tools Arm Embedded/9 2019-q4-major/lib/gcc/arm-none-eabi/9.2.1/include</header_search_path>
+                <header_search_path>C:/Program Files (x86)/GNU Tools Arm Embedded/9 2019-q4-major/arm-none-eabi/include/c++/9.2.1</header_search_path>
+                <header_search_path>C:/Program Files (x86)/GNU Tools Arm Embedded/9 2019-q4-major/arm-none-eabi/include/c++/9.2.1/arm-none-eabi</header_search_path>
+            </header_search_paths>
+            <indexed_header_paths>
+                <indexed_header_path>../build/libraries/base</indexed_header_path>
+                <indexed_header_path>../build/libraries/board/hardware</indexed_header_path>
+                <indexed_header_path>../build/libraries/drivers/gpio</indexed_header_path>
+                <indexed_header_path>../build/libraries/drivers/timer</indexed_header_path>
+                <indexed_header_path>../build/libraries/external</indexed_header_path>
+                <indexed_header_path>../build/libraries/motorController</indexed_header_path>
+                <indexed_header_path>../build/libraries/terminal</indexed_header_path>
+                <indexed_header_path>../build/src</indexed_header_path>
+                <indexed_header_path>../libraries/a4988/inc</indexed_header_path>
+                <indexed_header_path>../libraries/base</indexed_header_path>
+                <indexed_header_path>../libraries/board/hardware</indexed_header_path>
+                <indexed_header_path>../libraries/drivers/gpio</indexed_header_path>
+                <indexed_header_path>../libraries/drivers/timer</indexed_header_path>
+                <indexed_header_path>../libraries/drivers/uart/inc</indexed_header_path>
+                <indexed_header_path>../libraries/external</indexed_header_path>
+                <indexed_header_path>../libraries/motorController</indexed_header_path>
+                <indexed_header_path>../libraries/terminal</indexed_header_path>
+                <indexed_header_path>../platforms/mcu/EngModel</indexed_header_path>
+                <indexed_header_path>../src</indexed_header_path>
+            </indexed_header_paths>
+            <name>C/C++ from Compilation Database</name>
+            <pch_flags>
+                <use_compiler_flags>1</use_compiler_flags>
+            </pch_flags>
+            <status>enabled</status>
+            <type>C/C++ from Compilation Database</type>
+        </source_group_2fd9e6cd-1a40-4096-a649-56ef9c17a8c6>
+    </source_groups>
+    <version>8</version>
+</config>
